chore: make root component only export single comonent so hmr works
